+++ library/src/main/java/io/agora/board/fast/Fastboard.java
@@ -9,7 +9,7 @@ import io.agora.board.fast.model.FastRoomOptions;
 import io.agora.board.fast.model.FastStyle;
 
 public class Fastboard {
-    public static final String VERSION = "1.0.0";
+    public static final String VERSION = "1.0.1";
 
     private final FastboardView fastboardView;
     private FastStyle fastStyle;