+++ library/src/main/java/io/agora/board/fast/internal/FastOverlayManager.java
@@ -3,6 +3,7 @@ package io.agora.board.fast.internal;
 import android.util.SparseArray;
 
 import io.agora.board.fast.extension.FastVisiable;
+import io.agora.board.fast.extension.OverlayHandler;
 import io.agora.board.fast.extension.OverlayManager;
 import io.agora.board.fast.ui.OverlayLayout;
 
@@ -13,11 +14,13 @@ public class FastOverlayManager implements OverlayManager {
 
     private final SparseArray<FastVisiable> visiables = new SparseArray<>();
     private final OverlayLayout overlayLayer;
+    private final OverlayHandler overlayHandler;
 
     private int showKey = OverlayManager.KEY_NO_OVERLAY;
 
-    public FastOverlayManager(OverlayLayout overlayLayer) {
+    public FastOverlayManager(OverlayLayout overlayLayer, OverlayHandler overlayHandler) {
         this.overlayLayer = overlayLayer;
+        this.overlayHandler = overlayHandler;
         this.overlayLayer.hide();
         this.overlayLayer.setOnClickListener(v -> {
             hideAll();
@@ -82,8 +85,7 @@ public class FastOverlayManager implements OverlayManager {
             } else {
                 overlayLayer.show();
             }
-            // TODO
-            // fastContext.handleOverlayChanged(key);
+            overlayHandler.handleOverlayChanged(key);
         }
     }
 }