+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## [Version 1.1.0] - Pending
+
+- Fix: fix `ToolBox` selected state when overlay hide
+- New: add `FastRoomListener`.`onRoomPhaseChanged`
+
+## [Version 1.0.0] - 2021-03-20
+
+- Update: treat `FastRoom` as main apis class
+
 ## [Version 1.0.0-beta.5] - 2021-03-14
 
 - Update: update `FastRoomOptions.fastRegion` required